refactor file structures and apply naming conventions
It contains:

* put Layout and Header into components folder and make 'assets' folder lower case
* add providers folder
* move graphql generated types into types folder
* README additions regarding conventions
