+++ react/.prettierignore
@@ -1,2 +1,2 @@
-src/generated/
+src/types/generated/
 *.d.ts
