+++ react/src/App.tsx
@@ -1,14 +1,14 @@
 import 'regenerator-runtime/runtime';
 import React, { useContext, useEffect } from "react";
 import { Route, Routes } from "react-router-dom";
-import Boxes from "views/boxes/Boxes";
-import Locations from "views/locations/Locations";
-import BTLocation from "views/locations/BTLocation";
-import Layout from "Layout";
-import AutomaticBaseSwitcher from "views/automatic-base-switcher/AutomaticBaseSwitcher";
+import Boxes from "views/Boxes/Boxes";
+import BTLocations from "views/BTLocations/BTLocations";
+import BTLocation from "views/BTLocations/BTLocation";
+import Layout from "components/Layout";
+import AutomaticBaseSwitcher from "views/AutomaticBaseSwitcher/AutomaticBaseSwitcher";
 import { gql, useLazyQuery } from "@apollo/client";
-import { BasesQuery } from "generated/graphql";
-import { GlobalPreferencesContext } from "GlobalPreferencesProvider";
+import { BasesQuery } from "types/generated/graphql";
+import { GlobalPreferencesContext } from 'providers/GlobalPreferencesProvider';
 
 const useLoadAndSetAvailableBases = () => {
   const BASES_QUERY = gql`
@@ -50,7 +50,7 @@ const App = () => {
           <Route index element={<AutomaticBaseSwitcher />}></Route>
           <Route path=":baseId">
             <Route path="locations">
-              <Route index element={<Locations />} />
+              <Route index element={<BTLocations />} />
               <Route path=":locationId" element={<BTLocation />} />
             </Route>
             <Route path="boxes" element={<Boxes />} />
