+++ react/src/views/Boxes/Boxes.test.tsx
