+++ react/src/views/Boxes/components/GlobalFilter.tsx
