+++ react/src/views/Boxes/components/types.ts
@@ -1,0 +1,9 @@
+export type ProductRow = {
+    id: string;
+    name?: string;
+    gender?: string | null;
+    items: number;
+    size?: string | null;
+    state: string;
+  };
+
